internal/lcof/57.he-wei-sde-liang-ge-shu-zi/main_test.go
package main

import (
	"reflect"
	"testing"
)

type args struct {
	nums   []int
	target int
}

var tests = []struct {
	name string
	args args
	want []int
}{
	{"should be equal", args{[]int{2, 7, 11, 15}, 9}, []int{2, 7}},
	{"should be equal", args{[]int{10, 26, 30, 31, 47, 60}, 40}, []int{10, 30}},
}

func Test_twoSum(t *testing.T) {
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := twoSum(tt.args.nums, tt.args.target); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("twoSum() = %v, want %v", got, tt.want)
			}
		})
	}
}

func Test_twoSumFunc2(t *testing.T) {
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := twoSumFunc2(tt.args.nums, tt.args.target); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("twoSumFunc2() = %v, want %v", got, tt.want)
			}
		})
	}
}
